# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/listings": {
            "get": {
                "description": "Retrieves a paginated list of product listings for the authenticated user. Supports optional filtering by category and price range.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "Get product listings",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of listings per page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number to retrieve",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Category to filter listings",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Minimum price filter",
                        "name": "startprice",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Maximum price filter",
                        "name": "endprice",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "the person who created the listing",
                        "name": "creator",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of product listings",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - missing or invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/listings/": {
            "post": {
                "description": "This endpoint allows users to create new product listings.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "Create new listing",
                "parameters": [
                    {
                        "description": "Product details",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create listing response containing the product ID",
                        "schema": {
                            "$ref": "#/definitions/transport.CreateListingResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity - invalid JSON payload",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "This endpoint allows users to delete product listings.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "*/*"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "Delete a listing",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listing deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Product ID not found",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/listings/sold/bid": {
            "put": {
                "description": "This endpoint updates the sold status of a product based on a successful bid. The request body must contain the bid details in JSON format.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "Update sold status via bid",
                "parameters": [
                    {
                        "description": "Details of the bid used to mark the item as sold",
                        "name": "bid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SellItemViaBid"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item marked as sold successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity - invalid JSON payload",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/error.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "error.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "product_images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductImage"
                    }
                }
            }
        },
        "models.ProductImage": {
            "type": "object",
            "properties": {
                "base_64_image": {
                    "type": "string"
                },
                "checksum": {
                    "type": "string"
                },
                "compressedImage": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "order": {
                    "type": "integer"
                }
            }
        },
        "models.SellItemViaBid": {
            "type": "object",
            "properties": {
                "bid_id": {
                    "type": "string"
                },
                "item_id": {
                    "type": "string"
                }
            }
        },
...
        "transport.CreateListingResponse": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "string"
                }
            }
        }
    }
}